+++ packages/taro-platform-harmony-hybrid/src/api/apis/location/index.ts
@@ -1,3 +1,4 @@
 export * from './chooseLocation'
 export * from './getFuzzyLocation'
 export * from './getLocation'
+export * from './openLocation'
