+++ packages/taro-platform-harmony-hybrid/src/api/apis/taro.ts
@@ -222,6 +222,86 @@ function loadNavigationSytle () {
 
 loadNavigationSytle()
 
+// 设置位置选择样式
+function loadChooseLocationStyle () {
+  const css = `
+.taro_choose_location {
+  display: flex;
+  position: fixed;
+  top: 100%;
+  z-index: 1;
+  flex-direction: column;
+  width: 100%;
+  height: 100%;
+  background-color: #fff;
+  transition: ease top 0.3s;
+}
+.taro_choose_location_bar {
+  display: flex;
+  flex: 0 60px;
+  height: 60px;
+  background-color: #ededed;
+  color: #090909;
+  align-items: center;
+}
+.taro_choose_location_back {
+  position: relative;
+  flex: 0 40px;
+  margin-left: 10px;
+  width: 25px;
+  height: 30px;
+}
+.taro_choose_location_back::before {
+  display: block;
+  position: absolute;
+  left: 0;
+  top: 0;
+  border: solid 15px;
+  border-color: transparent #090909 transparent transparent;
+  width: 0;
+  height: 0;
+  content: "";
+}
+.taro_choose_location_back::after {
+  display: block;
+  position: absolute;
+  left: 3px;
+  top: 0;
+  border: solid 15px;
+  border-color: transparent #ededed transparent transparent;
+  width: 0;
+  height: 0;
+  content: "";
+}
+.taro_choose_location_title {
+  flex: 1;
+  padding-left: 30px;
+  line-height: 60px;
+}
+.taro_choose_location_submit {
+  margin-right: 25px;
+  padding: 0;
+  border: none;
+  width: 75px;
+  height: 40px;
+  background-color: #08bf62;
+  line-height: 40px;
+  font-size: 20px;
+  color: #fff;
+}
+.taro_choose_location_frame {
+  flex: 1;
+}
+`
+
+  const style = document.createElement('style')
+  style.innerHTML = css
+  document.getElementsByTagName('head')[0].appendChild(style)
+}
+
+loadChooseLocationStyle()
+
+
 taro.getApp = getApp
 taro.pxTransform = pxTransform
 taro.initPxTransform = initPxTransform
